backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4
python-dateutil==2.8.2
dnspython==2.4.2
email-validator==2.1.0.post1
python-dotenv==1.0.0
pydantic[email]==2.5.2
cryptography==41.0.5
aiofiles==23.2.1
python-magic==0.4.27
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
bcrypt==4.0.1
pyotp==2.9.0
qrcode==7.4.2
pillow==10.1.0
redis==5.0.1
celery==5.3.6
flower==2.0.1
prometheus-client==0.19.0
sentry-sdk==1.39.1 